Introduce the concept of scoped extractors
Refines the `FromSystem` trait, allowing implementors to define a type
as non-scoped, indicating that it performs read-access to the global
system state rather than a part of it indicated by the path linking the
job on the domain.

This also updates the handler and task definitions. An `is_scoped`
function has been added to handler to indicate that the handler uses
only scope extractors and is parallelizable.

Change-type: minor