Move testing helper functions into Worker
With the introduction of shared worker resources and the
`Res`extractor, testing helpers like `test_task` (previously in the domain module) or
`find_workflow` (previously in the planner module), no longer be
trusted to succeed unless in the context of a `Worker` instance.

This commit moves them under a `testing` module in worker that is
exported if `debug_assertions` is set.

Change-type: patch