Remove finished message from run_task logs
It is unnecessary since users can make use of the `close` event to
detect when the task terminated and it just causes noise on the logs.

It is different with the `starting` message, because tracing `new` event
doesn't convey enough information, and the `enter` event is emmited
every time the async task is polled by the tokio runtime, so it gets too
verbose.

Change-type: patch