Fix planner automated concurrency detection
This fixes how the planner chooses tasks that can be executed
concurrently from the list of candidates of each stage. Before it
would use the longest list of non conflicting paths, but that is wrong,
because a task for `/a` should be chosen over a task for `/a/b` because
of the candidate preference. The new calculation chooses the longest list
of non conflicting paths giving preference to shorter prefixes.

This also adds adds a new helper function for path conflict detection
when evaluating task methods and adds some testing for these helper
functions.

Change-type: patch