Improve conflict detection when planning
Methods are allowed to modify a different part of the state than the one
pointed by the path assigned to the method. This is quite useful and
powerful, because it allows to have methods that prepare the state to
perform a certain operation. An example is given in the added tests
where, a service creation needs to be preceeed by an image creation.

The problem with this is that we cannot just use the method path for
conflict detection but need to use the reported changes for each
sub-task to determine conflicts. This updates the planner to do that and
adds some tests.

Change-type: patch