Keep track of visited states during planning
This greatly reduces search time as it quickly discards plans that are
reordering of other plans and don't advance the search. This also
greatly reduces (and perhaps removes) the probability of infinite loops
during planning as there is no longer the posibility of jumping between
two visited states forever.

Change-type: patch