Use `State` trait for `Target` extractor
This now requires that the type used in the `Target` extractor
implements `State`, and uses the `::Target` type as the internal value
of the extractor. This results in more ergonomic code, particulary when
working with the `derive` macro, as it is no longer necessary to use the
target type, i.e `Target<ServiceTarget>` but instead `Target<Service>`
is used.

To extract targets for types that do not implement `State`, a new
`RawTarget` type is introduced, useful for passing targets to tasks from
a method.

This is a breaking change.

Change-type: minor