Abort Worker methods if the writer task closes
Previously a Worker with a closed writer task (due to an internal patch
error) could still receive `seek_with_interrupt` and `run_workflow`
requests. This is dangerous as the worker could be acting on an
inconsistent state and thus calculating an incorrect workflow.

This bug is a left-over from a previous version of the worker that would
consume the worker when seeking a target state.

Change-type: patch